[sil-combine] (projection (load)) -> (load (gep)), done right this time.
This patch canonicalizes projections of load to loads of gep so that passes like
the ARC optimizer and the load/store optimization in SILCodeMotion have a
canonical pattern to look for. As an example, often times in the stdlib one
will see something like this:

(retain (struct_extract (load)))

and,

(release (load (struct_element_addr)))

It does not matter which one we pick, just that we pick one.

Swift SVN r13324