[sil-aa] Initial commit just to set the AA API with alias stubs.
The idea is for AliasAnalysis to be a wrapper around a Cache of alias analysis
results that a pass creates when it needs to perform alias analysis queries in
an analogous manner to how passes use DominanceInfo.

When we get an "analysis" pass infrastructure, this should be made into an
analysis pass.

Currently the stubs just return AliasAnalysis::Result::MayAlias.

Swift SVN r13327