[utils] buildbot-script.sh: LLDB compatibility
The LLDB project requires an LLVM project containing both clang and
swift, while buildbot-script builds swift outside the LLVM tree.  The
obvious choice, if you don't want two copies of the source on your
system, is to symlink swift into llvm's tools directory.  Yes, building
lldb will still build llvm and swift again, but that's a problem for
another day.  To avoid building swift a /third/ time as part of swift's
LLVM build, we need to tell the buildbot-script to ignore that symlink
if it's there.

Swift SVN r13412