[Emacs support] Fix for first file opened
When visiting one of this project's C++ files for the first time in an
Emacs session, it wouldn't have loaded swift-project-settings.el in
time.  Setting c-file-style later in the list of directory locals fixes
this issue.  Also apply the same whitespace settings to objc-mode and
c-mode as c++-mode

Swift SVN r13632