Serialization: Don't throw away the conformances of ApplyInst substitutions.
We kinda need those. Limit the hack to pad it out with nulls to only apply in the case when the conformance list is empty but the original archetype requires conformances, which only occurs for archetypes or existentials.

At long last, we can build at -O0 again!

Swift SVN r13815