[tests] (Partially) Revive ASan build by updating XFAILs.
"opt" is not being built with ASan, so the IR-level ARC optimizer tests
won't work.

DebugInfo/variables-repl.swift seems to be passing, possibly because we're
not using the SwiftASTStreamerPass when we generate debug info anymore.

stdlib/NewString.swift is still failing; we're not seeing the expected
capacity for a particular string.

Swift SVN r13833