Added new benchmark Julia.
There are two versions, one that runs a single drawing iteration in a loop and
the other which runs a timer for 60 seconds, calling a method every 0.001
seconds.

This version I am committing now does a single drawing iteration.

Currently the status of the benchmark is the following:

When running with -O0 -g,

objc: 0.15 seconds.
swift: 5.85 seconds.

When running with -O3,

objc: 0.12 seconds.
swift: 4.65 seconds.

Swift SVN r13877