Generic dispatching prototype
Instead of flailing about trying to find the right combination of tricks
in the standard library code itself, create a separate test file.  This
demonstrates that we have everything we need to implement algorithm
specialization similar to what the C++ standard library does!

Next: moving on to apply the technique...

Swift SVN r13979