Remove duplicate code for handling mark_uninitialized.
checkAllocBoxUses had two places where mark_uninitialized was being
processed, the second one unreachable.

The code removed by this commit was the code added later, and it makes
the assumption that the mark_uninitialized is the only use of the value
being processed. Although that assumption might be reasonable and
currently true, it's not verified.

Swift SVN r14043