[driver] Pass -module-cache-path and -F to both compilation and module merging.
Previously we only passed -module-cache-path to the compilation jobs, which
meant the module merging jobs would default to using the system module cache
(which means double the Clang work if the module cache had been set, and
strange errors during development when Clang's serialized format changes).

We hadn't been passing -F to either job because we hadn't been worrying
about non-system frameworks, but that's likely to change soon.

I also consolidated the forwarding of -Xfrontend args in this commit.
I've convinced myself that this will pretty much always be necessary for the
merge-module step, and that we should have something else if we really
need to differentiate the two.

<rdar://problem/16113207>

Swift SVN r14114