[sil-dis] Initial commit of sil deserialization utility "sil-dis".
Currently supports dumping decls and SILModules from the standard library only.
This restriction is good enough for my purposes of verification of the
deserialization of SIL and allows me to ignore the complexity of having multiple
source files (which not being the stdlib implies most of the time).

Swift SVN r14187