Update SIL.rst with information about Typed Access TBAA.
This allows the optimizer to use the type information from loads, stores,
and address projections to disambiguate pointers. This will allow for
significantly more aggressive load/store forwarding of the “owning pointer”
pattern in the std lib by enabling the disambiguation of loads/stores of the
owning pointer that feed into retains/releases from loads/stores of the structs
data. This is a pervasive pattern seen in such important structs as Array,
String, and many others.

Swift SVN r14210