[test] Remove substitutions for binaries that will never move.
There's no reason to provide alternate versions of swift-ios-test,
swift-demangle, sil-extract, or lldb-moduleimport-test. There's no
reason to provide options for any of them either. And there's no
reason to warn about swift-ios-test being missing unless you're
building for ARM.

From now on, we should only be using "%foo" if:
A) 'foo' may have options we want to set for all invocations, or
B) we may want to run the tests with an alternate 'foo'.

Everywhere else, we should just use "foo". The build directory is already
in the path.

Swift SVN r14246