fix rdar://16119509 - <unknown>:0: error: super.init called multiple times in initializer
This is a problem whereby the parallel bitvector dataflow stuff in DI was
incorrectly caching speculated results in cycles.  In principle, this algorithm
should be reworked to use standard oldschool RPO bitwise dataflow algorithms, but
this patch is a minimal fix and the performance of this code is unlikely to be an issue
anyway.


Swift SVN r14322