Don't consider enum elements to be initializers.
Remove an odd enum hack we've had since before we had initializers, which
treated the cases of an enum as initializers that accept a value of
the type stored in the initializer. For example, given a typical
optional type definition like

  enum Optional<T> {
    case None
    case Some(T)
  }

One could directly construct an optional as, e.g.,
Optional<Int>(5). While this is the right thing for optional, it
doesn't make sense in the general case. Remove this behavior, and
explicitly add initializers in places where they make sense.


Swift SVN r14335