Remove root Makefile of Makefile build.
I'm delaying removing the rest of the Makefile build, as there
is a chance that somebody who is relying on the Makefile build
doesn't know it is going away.  If there is a reason to hold on
to it, it's easier just to add this one file back.

Swift SVN r14354