[build] Basic support for gyb
Now a "foo.bar.gyb" file in a list of sources gets automatically
processed through gyb to produce "foo.bar"

Renamed a simple .swift file in the stdlib (which needs a rename anyway)
to have a .gyb extension as a proof-of-concept.

To make this useful for our existing gyb files, we need to figure out
how to pass them the right arguments; that's coming next.

Swift SVN r14444