Fix a type checker crash on:
func braceStmt3() {
  {
    undefinedIdentifier {}
  }
}

caused by the parser trying to be helpful and unwrapping the closure's
BraceStmt.  Unfortunately, DeclContexts inside the BraceStmt are wrong with
this recovery approach.

Unfortunately, this causes us to produce some extra error messages that the
original recovery strategy tried to avoid.

One thing we could do alternatively without making QoI worse is trying to save
the parser position and trying to reparse with correct assumptions, but that
could slow down the parser, so I did not implement this.


Swift SVN r14456