Parser: don't set 'Invalid' bit on SubscriptDecls too eagerly.
This makes the typechecker skip the decl completely, and it does not set types
on parameter patterns (not even to ErrorType).  In fact, in the problematic
case the type checker can cope with the decl just fine, so don't set the
Invalid bit.

This fixes a crash in code completion when it encounters such a decl,
rdar://16132349


Swift SVN r14459