[utils] gyb: use % instead of @
It was hard to decide on a replacement character for @, which is a
particularly bad choice for gyb ever since we started using it for
attributes in Swift, since attributes often start a line.  Only
graphically-dense characters seem to look right.  '%' has the
disadvantage of clashing with SIL, where it begins many lines, but we're
not using gyb to generate any SIL files today, so I optimized for the
common case.

Swift SVN r14483